requirements.txt
Flask==3.1.3
twilio==9.10.4
python-dotenv==1.2.2
gunicorn==23.0.0